Breaking Types into a separate module.
